requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
distro==1.9.0
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
dotenv==0.9.9
et_xmlfile==2.0.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
joblib==1.5.1
lxml==5.3.0
mysql==0.0.3
mysqlclient==2.2.4
numpy==2.2.6
openai==1.83.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
playwright==1.52.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pyOpenSSL==25.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
qrcode==8.0
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.35.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==8.0
